requirements.txt
autoflake==1.3.1
certifi==2018.10.15
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.4.2
Django==2.1.5
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.9.0
entrypoints==0.3
flake8==3.7.8
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
mccabe==0.6.1
openapi-codec==1.3.2
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2018.7
requests==2.20.1
simplejson==3.16.0
uritemplate==3.0.0
urllib3==1.24.1
